Some bugs left

+++ EU_values/EU_values/Game/UI/Elements/Box.cs
@@ -8,7 +8,7 @@ public class Box : DrawableObject
     public Box(string name, float x, float y, float width, float height, Color color, bool? isFilled = true, bool? isCameraAffected = false) : base(name, x, y, width, height, color, isCameraAffected) {
         if (isFilled != null) IsFilled = isFilled.Value;
     }
-    public Box(string name, float x, float y, int width, int height, Image image, bool? isCameraAffected = false) : base(name, x, y, width, height, image, isCameraAffected) { }
+    public Box(string name, float x, float y, float width, float height, Image image, bool? isCameraAffected = false) : base(name, x, y, width, height, image, isCameraAffected) { }
 
     public override void Update(float timeDelta, int timeRemaining, int timeDifference) { }
 }
